+++ libsolidity/formal/SMTEncoder.cpp
@@ -1098,13 +1098,18 @@ bool SMTEncoder::shouldAnalyze(SourceUnit const& _source) const
 		m_settings.contracts.has(*_source.annotation().path);
 }
 
+bool SMTEncoder::skipContract(ContractDefinition const& _contract) const
+{
+	return !_contract.canBeDeployed();
+}
+
 bool SMTEncoder::shouldAnalyze(ContractDefinition const& _contract) const
 {
-	if (!_contract.canBeDeployed())
+	if (skipContract(_contract))
 		return false;
 
 	return m_settings.contracts.isDefault() ||
-		m_settings.contracts.has(_contract.sourceUnitName());
+		m_settings.contracts.has(_contract.sourceUnitName(), _contract.name());
 }
 
 void SMTEncoder::visitTypeConversion(FunctionCall const& _funCall)