+++ libsolidity/formal/SMTEncoder.h
@@ -238,6 +238,7 @@ protected:
 	void visitFunctionIdentifier(Identifier const& _identifier);
 	virtual void visitPublicGetter(FunctionCall const& _funCall);
 
+	bool skipContract(ContractDefinition const& _contract) const;
 	/// @returns true if @param _contract is set for analysis in the settings
 	/// and it is not abstract.
 	bool shouldAnalyze(ContractDefinition const& _contract) const;