+++ test/libsolidity/smtCheckerTests/verification_target/only_selected_contracts_ignore_target_in_public_function.sol
@@ -1,0 +1,12 @@
+contract B {
+   function fail() pure public { assert(false); }
+}
+
+contract A {
+    function fail() pure public { assert(false); }
+}
+// ====
+// SMTContract: A
+// SMTEngine: chc
+// ----
+// Warning 6328: (113-126): CHC: Assertion violation happens here.\nCounterexample:\n\n\nTransaction trace:\nA.constructor()\nA.fail()